Support custom annotations
Now it is possible to get access to the annotations of an Ingress
resource from the template for Ingress resources. This allows
users to implement custom annotations by modifying the template
to insert NGINX configuration based on the presence on an
annotation or its value.

Additionally, the Ingress name and namespace is also avaiable
in the template.