Fix AppProtect tests
Because of the changes in 70e6b8c1, the Ingress Controller will not
become ready until all informers are synced.

An informer for a CR will not sync until the corresponding CRD
exists in the cluster.

This commit ensures that all required CRDs are created prior to
deploying the IC, so that CR informers can sync.