GCP Marketplace integration for NGINX Ingress
Google Cloud's marketplace requires that the application report usage
metrics periodically to GCP's service endpoint. UBB Agent SDK is used
to send a count of 60 to the endpoint every 60 seconds, using credentials
added as a secret during marketplace provisioning. If the secret is
unavailable or invalid, ingress controller will terminate after five
successive failures.

Runtime requirements:

 - The embedded variable gcpServiceName must match the fully-qualified
   GCP service identifier, as established during product listing
 - A per-deployment Secret containing a customer identifier and JSON
   key that are used to authenticate to the usage-based billing
   endpoint

GCP Marketplace is click-to-deploy only; in addition to ingress controller,
a deployer and tester container must be provided as a package. The source
code for these containers is in a separate repo.