Update InternalRoute server_name
In order for NGINX Service Mesh to support multiple egress endpoints, the server block for the internal server route cannot use the pod name as the server_name. Multiple egress endpoints are grouped under the same service account, so this is the SpiffeID DNS name that must be used, specifically of the format <serviceaccount>.<namespace>.svc.