skip optional external-status-address config if unset
We are currently unconditionally setting a value for
external-status-address from the config map. If it does not exist an
empty value is saved. This fills the cached status with an unset (empty)
value.

This empty value would then be written to the ingress object eventually.
Propagating this empty value can lead to all kind of failures.

We have a guard in place that an uninitialized status cache will not be
written to the ingress object status. So to avoid writing empty
values to the ingress status let us check whether the configuration for
external-status-address exists.