feat: Fix instruction link for NGINX Ingress Helm Operator
This commit makes a small change to the the third prerequisite in
the installation document to point at the correct instructions
it is referencing.

It also expands the link itself to make it clearer what precisely
the reader is being referenced to, and prefixes F5 to the first
instance of the product name (Ignoring the ephemeral note).