fix: Add missing directory to NGINX App Protect DoS Instructions
This miniature commit changes the path for the CD command in the second
prerequisite of the "Installation with NGINX App Protect DoS" document.
A reader with existing knowledge would spot the issue immediately, but
one without might be left confused why the subsequent commands fail.

Additional, this commit fixes some minor noun issues for the
documentation based on the product naming conventions.