Use the "runtime default" seccomp profile
seccomp profiles allow sandboxing processes, in particular to restrict
allowed syscalls from applications to the kernel. Kubernetes default in
current release is Unconfined seccomp profile, which is essentially
privileged. It is preferred for security purposes to restrict this.

KEP-2413 proposes that RuntimeDefault will become the new default for
Kubernetes. With Kubernetes v1.25, this is in Beta, and available with
`SeccompDefault` feature gate and `--seccomp-default` CLI flag.

`nginx-ingress` should switch to this new default, in order to ensure
compatibility down the line, as well as enable enhanced security on
older Kubernetes versions.

Co-authored-by: Christian Ihle <blurpy@users.noreply.github.com>