Restructure and rewrite parts of the Docker JWT Token Secret Page
This commit makes some large changes in the layout, gramma and overall
phrasing of this installation guide. Many instances of the NGINX Ingress
Controller's name were changed, and where possible, text was made more
concise. Additionally, many links were added or renamed for explicit
ease in finding or understanding the context related to an instruction.