Restructure, rename and rewrite various troubleshooting documents
This commit adds the changes I have made thus far to the troubleshooting
section of NGINX Ingress Controller's documentation: there is more to be
done yet, but I am pushing these changes for others to see.

The product naming is inconsistent, as is the layout of the pages -
often, appropriate markdown is not used where it would make sense to,
and formatting is otherwise missing.

I will continue to update the documentation to fix these instances.