Re-add changes to  JWT & Docket documentation
This branch has had multiple merge conflicts and force pushes,
completely erasing a number of large changes to the page to the point of
unintentionally removing entire sections.

This commit takes the last known good working version and adds the
changes back as a fresh new point for the branch.