Fix product and command nouns, add tabs and adjusting phrasing
This commit adds a handful of changes to the Helm upgrade installation
guide for NGINX Ingress Controller. It primarily changes the body of
text to favour plain English over technical jargon where possible, and
also recontextualises it to be less negatively-framed.

It also adds some standardised formatting for NGINX products such as the
use of note callouts and tabs, and reduces the inline comments to using
single backticks instead of triples, which are only necessary for large
blocks. The list numeration is also changed to the automatic style.