Change links on NGINX Plus introduction page to relative links
This commit fixes the reported issue in #4208 and changes other links on
the same page to follow convention. Prior to this commit, the page had a
combination of linking to the live, production URLs for other
documentation while adding the markdown extension.

Instead, the links now use the hugo relref (relative reference)
convention for consistency, and have dropped the filename extension,
fixing the links and giving them more consistent development behaviour.