Ensure Debian Plus starts assigning users from 101
When running `adduser` or `addgroup`, adduser.conf controls various
configuration flags. Notably for us, there is `FIRST_SYSTEM_UID` and
`FIRST_SYSTEM_GID` which determine what UIDs and GIDs are assigned
for system users.

The Nginx images are historically built with UID 101 for `nginx` user.
However, Debian 12 cleaned up its image and UID 100 became available.
By default, the range `100-999` is used for system users.
That is bad for the Nginx image as UID 101 is expected.

This patch simply ensures the first ID used for package install is 101.