Rewrite NGINX App Protect WAF Integration Section (#4726)
This commit rewrites the installation document for NGINX App Protect WAF
to match the format posited in the prior NGINX App Protect DOS rewrite.
As part of the changes, a snippet of text indicating the deployment options
for NGINX Ingress Controller was turned into an include.

This commit adds a new include file for custom resources, and
subsequently adds it to the relevant App Protect WAF/DoS documentation.

A few documentation improvements are also included to rephrase sections of
the configuration document for conciseness and clarity.