Modify documentation metadata and improve opening paragraphs
This commit removes all unnecessary metadata from the frontmatter of
every documentation markdown file. It also moves the description data
from the frontmatter to the opening section or removes duplicate
information on a case by case basis.

These changes are made in advance of further efforts to improve the
documentation contents as well as the tooling for the documentation.