NAP link fixes & improvements to troubleshooting documentation (#5605)
This commit updates broken NAP links to include the v4 path and changes
the location of the troubleshooting documents so that they are colocated with
the other NAP documentation.

As part of this effort a number of improvements were made to the WAF, DoS
and troubleshooting documents related to newer documentation standards.
These changes include the use of title case and excision of gerunds.