Add F5 prefix to configuration opening paragraphs, style consistency
This commit changes each of the opening paragraphs of the Configuration
section to ensure they have the F5 prefixed name for the first instance
of NGINX Ingress Controller.

It also includes a large amount of other changes:

- Relative links
- Erroneous product names ("the Ingress Controller")
- Contemporary table formatting
- Updated and decoupled input/output code blocks
- Sentence case for page titles and headings
- Small contextual rewrites
- Updated links to appropriate tags

A small number of unrelated changes were also included, regarding NGINX
App Protect and a small typo in a command line option.