Documentation style changes for the "Installation" section (#6015)
This commit is one of a series to update the documentation set to
adhere to NGINX's contemporary documentation standards.

The three most prominent changes are prepending F5 to the first instance
of an NGINX product, changing titles to sentence case, and removing gerunds.

A number of pages also had their filenames changed to
reflect their contents and to create shorter URL slugs.

---------

Signed-off-by: Alan Dooley <a.dooley@f5.com>
Co-authored-by: Shaun <s.odonovan@f5.com>
Co-authored-by: Venktesh Shivam Patel <ve.patel@f5.com>