Update documentation Makefile and README
This commit updates the documentation Makefile and README, reflecting
changes to it across NGINX's open source repositories for consistency.
It removes unnecessary targets: Netlify is no longer used, nor was the
Makefile used for Netlify for quite some time.

The guidance for Makefile targets in the README is updated accordingly,
as well formatting instructions for how we use Hugo contemporaneously.