Migrate access control example to documentation website
This commit migrates the access control example to the documentation
website, converting the existing README into a page formatted for the
production website. It also cleans up the frontmatter of the surrounding
documents so that pages can be more easily re-arranged in the future.

The converted page make use of single source embedded code blocks using
files in the GitHub repository, as well as a highlighting feature of
code blocks that has not been used in previous documentation.