Update build instruction link, remove EoS versions from specifications
This commit updates the README in the build subfolder to point to the
latest version of the build documentation, as the link was updated as
part of previous documentation changes.

It also removes the last two entries from the supported Kubernetes
versions table of the Technical specifications page, as they are now out
of software and technical support.