Fix mountPath directory for NAP5 policy documentation
This commit updates the path for the volumeMounts shown in the NAP5
compile documentation to use the newest convention, and also adds a
missing parameter from the `kubectl cp` example at the end of the page.

Issue reported by Fabrizio Fiorucci, as it can cause confusion and
possible error for a reader following along and reading the NAP
documentation concurrently.